+++ lang/src/lib.rs
@@ -65,7 +65,7 @@ pub use solana_program;
 pub use anchor_attribute_event::{emit_cpi, event_cpi};
 
 #[cfg(feature = "idl-build")]
-pub use anchor_syn::{self, idl::build::IdlBuild};
+pub use idl::IdlBuild;
 
 #[cfg(feature = "interface-instructions")]
 pub use anchor_attribute_program::interface;
@@ -423,7 +423,7 @@ pub mod prelude {
     pub use super::{emit_cpi, event_cpi};
 
     #[cfg(feature = "idl-build")]
-    pub use super::IdlBuild;
+    pub use super::idl::IdlBuild;
 
     #[cfg(feature = "interface-instructions")]
     pub use super::interface;