+++ lang/syn/src/codegen/program/mod.rs
@@ -39,7 +39,7 @@ pub fn generate(program: &Program) -> proc_macro2::TokenStream {
 
     #[cfg(feature = "idl-build")]
     {
-        let idl_build_impl = crate::idl::build::gen_idl_print_fn_program(program);
+        let idl_build_impl = crate::idl::gen_idl_print_fn_program(program);
         return quote! {
             #ret
             #idl_build_impl